Move apolloClientWrapper to utils file.
